feat(compute): wire ZK circuits to keeper via CircuitManager
- Add CircuitManager that imports and manages circuits from circuits/ package
- Integrate ComputeCircuit, EscrowCircuit, and ResultCircuit (v2 circuits)
- Add lazy initialization for circuit compilation (~30-60s operation)
- Add keeper methods: GetCircuitManager, InitializeCircuits
- Add helper methods for each circuit type verification
- Update placeholder comment in verification.go to reference implementation
- Circuits now properly imported and used instead of sitting unused

The ZK system now provides:
- Compute verification: ~40k constraints, 4 public inputs
- Escrow release: ~26k constraints, 5 public inputs
- Result correctness: ~45k constraints, 4 public inputs

+++ tests/benchmarks/compute_bench_test.go
@@ -164,9 +164,9 @@ func BenchmarkProviderSelection(b *testing.B) {
 			"provider-"+string(rune('0'+i)),
 			"http://provider:8080",
 			types.ComputeSpec{
-				CpuCores:       uint32(4 + i),
-				MemoryMb:       uint32(8192 + i*1024),
-				StorageGb:      uint32(100 + i*10),
+				CpuCores:       uint64(4 + i),
+				MemoryMb:       uint64(8192 + i*1024),
+				StorageGb:      uint64(100 + i*10),
 				GpuCount:       0,
 				GpuType:        "",
 				TimeoutSeconds: 3600,
